This is SVG reader, after Amir made changes to it. Ready for PR
