+++ Moush_iOS/Moush_iOS/ContentView.swift
@@ -1,38 +1,10 @@
-//
-//  ContentView.swift
-//  Moush_iOS
-//
-//  Created by The Odd Institute on 2023-04-30.
-//
-
 import SwiftUI
 
-struct ContentView: View {
+struct ContentView: View
+{
     var body: some View
     {
-        NavigationStack
-        {
-            NavigationLink {
-                SwiftUI_SVG()
-            } label: {
-                Text("SVG Mercedes")
-            }
-            
-            NavigationLink {
-                SVGView()
-            } label: {
-                Text("SVG og")
-            }
-
-            VStack
-            {
-                Image(systemName: "globe")
-                    .imageScale(.large)
-                    .foregroundColor(.accentColor)
-                Text("Hello, world!")
-            }
-            .padding()
-        }
+        EditSvgScreen()
     }
 }
