+++ Moush_iOS/Moush_iOS/SvgLayersView.swift
@@ -1,0 +1,29 @@
+import SwiftUI
+
+struct SvgLayersView: View
+{
+    @ObservedObject
+    var vm: ViewModel
+    
+    
+    var body: some View
+    {
+        List
+        {
+            ForEach(vm.paths.indices, id: \.self) { i in
+                LayerCellView(pathModel: $vm.paths[i])
+                    .listRowSeparatorTint(.black)
+                    .listRowInsets(EdgeInsets())
+                    .onTapGesture {
+                        vm.setSelectedPathIndex(i)
+                    }
+            }
+        }
+    }
+}
+
+struct SvgLayersView_Previews: PreviewProvider {
+    static var previews: some View {
+        SvgLayersView(vm: ViewModel())
+    }
+}
